Adding Exploratory Data Analysis Project
This project was a collaborative final for a graduate level Regressions Analysis course I completed during my undergraduate studies. For this project my group was tasked with finding, cleaning, and analyzing a dataset as well as completing an analysis project. Our project had to do with exploring the prospect of Multinomial Logistic Regression for classifying Spotify song data.